01_Setup/requirements.txt
# Notion API SDK
notion-client==2.2.1

# Environment variable management
python-dotenv==1.0.0

# HTTP requests for file uploads
requests==2.31.0
